requirements.txt
tiktoken~=0.12.0
requests~=2.33.1
numpy~=2.4.4
scikit-learn~=1.8.0
sentence-transformers~=5.4.0
pandas~=3.0.2
pypdf2~=3.0.1
python-docx~=1.2.0